fixed re-render on task restore + init settings page & route
