+++ client/src/routes/app-notifications/notification.css
@@ -8,6 +8,13 @@
   width: 250px;
   display: flex;
   flex-direction: column;
-  justify-content: flex-start;
-  align-items: space-around;
+  justify-content: safe;
+  align-items: flex-start;
+}
+
+.notification-container button {
+  outline: none;
+  border: 1px solid;
+  border-radius: 5px;
+  padding: 0.1em 0.2em;
 }
