+++ client/src/routes/project-routes/project.tsx
@@ -17,22 +17,16 @@ import {
   restoreTask,
   setFavorite,
 } from "../../dummyDB";
-import {
-  Projects,
-  Task,
-  TaskStatus,
-  Todo,
-  TodoStatus,
-} from "../../types/project";
+import { Projects, Task } from "../../types/project";
 import EditSVG from "../../images/Edit.svg";
 import DeleteSVG from "../../images/Delete.svg";
 import FavSVG from "../../images/Star_filled.svg";
 import UnFavSVG from "../../images/Star_blank.svg";
 import AddSVG from "../../images/Add.svg";
-import DoneSVG from "../../images/Checkmark.svg";
 import React, { useEffect, useState } from "react";
 import "../task-routes/tasks.css";
 import ActionNotifcation from "../app-notifications/action-notifcation";
+import TaskCard from "../task-routes/task-card";
 
 export const loader = async ({ params }: { params: Params<string> }) => {
   const id = params.projectId;
@@ -78,11 +72,7 @@ export const action = async ({
     return { removed_task, key };
   } else if (formData.markTodo) {
     const todoLocation = formData.markTodo.toString().split(",");
-    const res = await markTodo(
-      id,
-      parseInt(todoLocation[0]),
-      parseInt(todoLocation[1])
-    );
+    await markTodo(id, parseInt(todoLocation[0]), parseInt(todoLocation[1]));
   }
 };
 
@@ -90,18 +80,17 @@ const Project = () => {
   const { project }: { project: Projects } = useLoaderData();
   const [isFav, setFav] = useState(project.favorite ? true : false);
   const [showNotification, setShowNotification] = useState(false);
-  const {
-    removed_task,
-    key,
-  }: { removed_task: Task | undefined; key: number | undefined } =
-    useActionData() ?? {};
+  const recent_delete: {
+    removed_task: Task | undefined;
+    key: number | undefined;
+  } = useActionData() ?? {};
   const [deletedTasks, setDeletedTasks] = useState<
     { removed_task: Task; key: number }[]
   >([]);
 
   useEffect(() => {
-    if (removed_task && key) {
-      deletedTasks.push({ removed_task, key });
+    if (recent_delete.removed_task && recent_delete.key) {
+      deletedTasks.push(recent_delete);
       setShowNotification(true);
     }
     const removeNotification = setTimeout(() => {
@@ -111,7 +100,7 @@ const Project = () => {
       setShowNotification(false);
       clearTimeout(removeNotification);
     };
-  }, [removed_task]);
+  }, [recent_delete, deletedTasks]);
 
   async function restoreLastDeletedTask(task: Task, index: number) {
     project.tasks.splice(index, 0, task);
@@ -120,6 +109,7 @@ const Project = () => {
       state.pop();
       return state;
     });
+    setShowNotification(false);
   }
 
   return (
@@ -175,7 +165,8 @@ const Project = () => {
         <ActionNotifcation
           content={{ message: "Task Deleted" }}
           action={{
-            name: "Undo",
+            name: "Undo delete",
+            target: "Task",
             execute: () => {
               const { removed_task, key } = deletedTasks.splice(-1)[0];
               restoreLastDeletedTask(removed_task, key);
@@ -213,132 +204,5 @@ const Tasks = ({ tasks }: { tasks: Task[] | undefined }) => {
         )}
       </div>
     </div>
-  );
-};
-
-const TaskCard = ({ task, index }: { task: Task; index: number }) => {
-  const [magicStyle, setMagicStyle] = useState("magictime swashIn");
-  const [gridRow, setGridRow] = useState("");
-  const [hasCompletedTodos, setHasCompletedTodos] = useState(false);
-  // change the span of task cards depending on their length
-  useEffect(() => {
-    const this_card = document.querySelectorAll(".task-card")[index];
-    const height = parseInt(getComputedStyle(this_card).height);
-    const span_ratio = parseInt((height / 100).toString());
-    setGridRow(`span ${span_ratio > 2 ? span_ratio - 1 : span_ratio}`);
-  }, [task, index]);
-
-  useEffect(() => {
-    let completed_todos = task.todos.filter(
-      (todo) => todo.status === TodoStatus.DONE
-    );
-    completed_todos.length
-      ? setHasCompletedTodos(true)
-      : setHasCompletedTodos(false);
-  }, [task.todos]);
-
-  return (
-    <div
-      className={`task-card ${magicStyle}`}
-      key={index}
-      style={{
-        gridRow: gridRow,
-      }}
-    >
-      <h2>{task.name}</h2>
-      <div>{task.deadline?.toLocaleString()}</div>
-      <ul className="todos">
-        {task.todos && task.todos.length ? (
-          task.todos.map((todo, key) => {
-            if (todo.status === TodoStatus.AWAITING) {
-              return (
-                <TodoComponent
-                  todo={todo}
-                  key={key}
-                  index={key}
-                  parent={index}
-                />
-              );
-            }
-          })
-        ) : (
-          <em>No awaiting Todos.</em>
-        )}
-
-        {hasCompletedTodos && <h5>Completed Todos: </h5>}
-        {task.todos && task.todos.length ? (
-          task.todos.map((todo, key) => {
-            if (todo.status === TodoStatus.DONE) {
-              return (
-                <TodoComponent
-                  todo={todo}
-                  key={key}
-                  index={key}
-                  parent={index}
-                />
-              );
-            }
-            return null;
-          })
-        ) : (
-          <></>
-        )}
-      </ul>
-      <Form method="post" className="task-actions">
-        <button
-          type="submit"
-          name="editTask"
-          value={index}
-          className="edit-task-btn"
-        >
-          <img src={EditSVG} />
-        </button>
-        <button
-          type="submit"
-          name="deleteTask"
-          value={index}
-          className="delete-task-btn"
-          onClick={() => {
-            setMagicStyle("magictime holeOut");
-            setTimeout(() => {
-              setMagicStyle("magictime");
-            }, 200);
-          }}
-        >
-          <img src={DeleteSVG} />
-        </button>
-      </Form>
-    </div>
-  );
-};
-
-const TodoComponent = ({
-  todo,
-  index,
-  parent,
-}: {
-  todo: Todo;
-  index: number;
-  parent: number;
-}) => {
-  return (
-    <li>
-      <span>{todo.content}</span>
-      <span>
-        {todo.status === TodoStatus.AWAITING ? (
-          <Form method="post">
-            <button
-              type="submit"
-              name="markTodo"
-              title="Mark as done"
-              value={[parent.toString(), index.toString()]}
-              className="mark-todo-done-btn"
-            >
-              <img src={DoneSVG} />
-            </button>
-          </Form>
-        ) : null}
-      </span>
-    </li>
   );
 };
