+++ client/src/routes/user-settings.tsx
@@ -1,0 +1,13 @@
+import { useState } from "react";
+
+const SettingsPage = () => {
+  const [user_settings, setUserSettings] = useState({
+    theme: "dark_mode",
+    offline_mode: true,
+    username: "",
+    email: "",
+  });
+  return <>Settings Incoming...</>;
+};
+
+export default SettingsPage;
